+++ example-app/controllers/auth.js
@@ -14,7 +14,7 @@ const httpOnlyCookie = `HttpOnly;Path=/;SameSite=${COOKIE_SAMESITE};` + (COOKIE_
 
 const checkGithub = async (req, res) => {
   try {
-    const { code, state } = req.body
+    const { code, state } = req.query
     const { data } = await axios.post('https://github.com/login/oauth/access_token', {
       client_id: GITHUB_CLIENT_ID, client_secret: GITHUB_CLIENT_SECRET, code, state
     }, {
@@ -31,7 +31,8 @@ const checkGithub = async (req, res) => {
     // SameSite=None; must use with Secure;
     // may need to restart browser, TBD set Max-Age, ALTERNATE use res.cookie, Signed?
     if (COOKIE_HTTPONLY) res.setHeader('Set-Cookie', [`token=${tokens.token};`+ httpOnlyCookie])
-    return res.status(200).json(tokens)
+    // return res.status(200).json(tokens)
+    return res.redirect('http://127.0.0.1:3000/native/index.html' + '#' + tokens.token) // use url fragment... // TBD make callback URL configurable
   } catch (e) {
     console.log('github auth err', e.toString())
   }