+++ example-native/router.js
@@ -5,7 +5,7 @@ import SignIn from './views/signin.js'
 
 const { createRouter, createWebHistory } = VueRouter
 
-const routerHistory = createWebHistory('/example-native/')
+const routerHistory = createWebHistory('/native/')
 
 const AuthGuard = async (to, from, next) => {
   // console.log('AuthGuard', store)
@@ -40,13 +40,7 @@ const router = createRouter({
     { meta: { requiresAuth: true, layout: 'layout-secure' }, beforeEnter: AuthGuard, path: '/ui1', component: () => import('./views/ui1.js'), name: 'ui1' },
     { meta: { requiresAuth: true, layout: 'layout-secure' }, beforeEnter: AuthGuard, path: '/ui2', component: () => import('./views/ui2.js'), name: 'ui2' },
     { meta: { requiresAuth: true, layout: 'layout-secure' }, beforeEnter: AuthGuard, path: '/ui3', component: () => import('./views/ui3.js'), name: 'ui3' },
-    {
-      meta: { requiresAuth: true, layout: 'layout-secure' },
-      beforeEnter: AuthGuard,
-      path: '/admin',
-      component: Admin, // () => import('./views/about.js') // TBD use lazy loading
-      name: 'admin'
-    },
+    { meta: { requiresAuth: true, layout: 'layout-secure' }, beforeEnter: AuthGuard, path: '/admin', component: Admin, name: 'admin' },
     {
       meta: { requiresAuth: false, layout: 'layout-public' },
       path: '/',