+++ example-webpack/src/layouts/Admin.vue
@@ -58,8 +58,7 @@ export default {
         { icon: 'dashboard', title: 'Dashboard', link: '/dashboard' },
         { icon: 'list_alt', title: 'Authors', link: '/authors' },
         { icon: 'list_alt', title: 'Categories', link: '/categories' },
-        { icon: 'list_alt', title: 'Books', link: '/books' },
-        { icon: 'list_alt', title: 'Test Stuff', link: '/test', groups: ['TestGroup'] }
+        { icon: 'list_alt', title: 'Books', link: '/books' }
       ]
     }
   },