+++ example-webpack/src/pages/Dashboard.vue
@@ -9,8 +9,6 @@
                 <v-flex xs12>
                   <h2 class="py-1 text-center">Test Your Other Stuff Here</h2>
                   <p class="py-1 text-center">Click on the top left menu icon to navigate</p>
-                  <p class="py-1 text-center">On this page, websocket messages are sent to server every 10 seconds (websocket must be enabled .env file). Message will appear server console.log</p>
-                  <p class="py-1 text-center">Message to server will be echoed back here... ({{ wsMsg }})</p>
                   <p class="py-1 text-center">
                     <v-btn @click="testGraphQL">Test GraphQL</v-btn>
                     <v-btn @click="testAuth">Test Auth</v-btn>
@@ -42,37 +40,14 @@ export default {
   data () {
     return {
       hello: 'NA',
-      wsMsg: 'No WS Message Yet',
       testResult: 'NA'
     }
   },
   created () {
-    this.priceTimerId = null
   },
   beforeDestroy () {
-    if (this.$socket) {
-      delete this.$options.sockets.onmessage
-      clearInterval(this.priceTimerId)
-    }
   },
   async mounted () {
-    if (this.$socket) {
-      this.$options.sockets.onmessage = (rv) => {
-        // console.log('onmessage', rv) // create listener
-        const data = JSON.parse(rv.data)
-        // console.log('onmessage', typeof JSON.parse(rv.data))
-        this.wsMsg = data.args
-      }
-      // this.$store.state.user.token
-      this.priceTimerId = setInterval(async () => {
-        this.$socket.sendObj({
-          call: 'msg-from-client',
-          args: 'Hello from client! ' + parseInt(Date.now() / 1000)
-        })
-        if (navigator.onLine) this.updateNetworkError(false)
-        else this.updateNetworkError(true)
-      }, 10000)
-    }
   },
   methods: {
     async testHealth () {
@@ -119,9 +94,6 @@ export default {
       const rv = await apolloClient.query({ query: gql`query DoHello($message: String!) { hello(message: $message) }`, variables: { message: 'Meow' } })
       console.log('testGraphQL', rv.data.hello)
       this.hello = rv.data.hello
-    },
-    updateNetworkError (flag) {
-      this.$store.dispatch('setNetworkError', flag)
     }
   }
 }